requirements.txt
streamlit==1.30.0
newspaper3k==0.2.8
gTTS==2.3.2
lxml[html_clean]==4.9.3
rich==14.2.0
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2
